fix warnings

+++ mckrnl/driver/timer/hpet.c
@@ -29,7 +29,7 @@ char* hpet_get_device_name(driver_t* driver) {
 void hpet_init(driver_t* driver) {    
     hpet_table_t* hpet_table = (hpet_table_t*) find_SDT("HPET");
     // vmm_map_page(kernel_context, (uintptr_t) hpet_table, (uintptr_t) hpet_table, PTE_PRESENT | PTE_WRITE);
-    hpet_t* hpet = (hpet_t*) hpet_table->address;
+    hpet_t* hpet = (hpet_t*) (uint32_t) hpet_table->address;
     vmm_map_page(kernel_context, (uintptr_t) hpet, (uintptr_t) hpet, PTE_PRESENT | PTE_WRITE);
 
     driver->driver_specific_data = hpet;
